feat(pager,lock): windows mmap — placeholder-reservation data mapping, section-mapped lock file
Chunk 6 of the platform-ports plan; the user selected the
placeholder-reservation collapse. Windows sections cannot express the
unix contract natively (a read-only section cannot exceed the file;
growth does not extend views), so the data mapping is a VirtualAlloc2
PLACEHOLDER reservation of MaxSize whose base never moves, with the
file-backed extent mapped in as PAGE_READONLY MapViewOfFile3 views:

- GROWTH maps the new extent into the placeholder (existing views and
  borrowed page pointers untouched). A peer's growth is covered where
  file-size knowledge refreshes: attachState (right after its fresh
  stat, in the fallible zone) and per-snapshot reader mapping
  creation.
- SHRINK unmaps tail views first (windows refuses SetEndOfFile under
  a mapped view), truncates, remaps a straddling view's surviving
  prefix. Cross-process: SetEndOfFile fails while ANY peer maps the
  file — multi-process shrink is structurally DEFERRED on windows
  (spec'd); the failed truncation restores coverage and a later
  commit retries.
- Fragmentation invariant: [coverage, reservation) is always ONE
  placeholder — every exit (mapExtent success/failure at either
  stage, unmapFrom, munmap) re-coalesces, with structural placeholder
  counting so the undocumented single-placeholder coalesce is never
  invoked.
- Failure posture: any post-unmap shrink failure restores coverage to
  the extent the file actually has (pre-truncate vs truncated,
  per-site); an unrestorable tear panics — the commit is already
  durable and continuing guarantees an arbitrary later access
  violation.
- Alignment regime: windows file lengths kept 64 KiB-aligned
  (platformTruncate rounds up; Init routed through it; a foreign
  unix-created file is aligned once at writable open; read-only open
  of an unaligned file fails — documented degradation). Page
  accounting keeps un-rounded sizes; slack sits above HighWaterMark.
- Old-windows degradation: the three kernelbase procs are
  pre-flighted; pre-1803 gets a clean versioned error, not a
  LazyProc panic. Build tags pin windows && (amd64 || arm64) — the
  spec's arch set; windows/386 compiles to the erroring stub.

The lock file needs none of this: fixed size, never resized — one
PAGE_READWRITE section + FILE_MAP_READ|WRITE view; distinct
per-process sections over one local file share physical pages, so
the atomic slot protocol stays cross-process coherent.

mmap-strategy.md gains §Windows (the collapse, coverage rules,
deferral semantics, alignment regime, degradations); the read-only
handle's pager comment now states raw reads go through per-snapshot
pagers (its coverage is open-time-fixed on windows).

Triage at 6.1: 0 new match.

Spec-first: canonical=mmap-strategy.md §Invariants + §mmap Resizing;
contract=reservation sized to MaxSize, stable base, never remapped,
fault-beyond-file; mechanism=placeholder collapse (§Windows, settled
this chunk with the user's selection); collapse-check=faithful
(stable base and pointer validity preserved; deferred multi-process
shrink and the alignment slack are documented degradations, not
contract changes).

Mutation posture: windows-only code unexecutable on linux — attested;
executor is the windows CI leg (next chunks). Linux delta (hook
no-ops, osFileOps.Truncate indirection, shrink-closure restructure)
covered by the full -race suite, green.

Adversarial review: 4 rounds, converged; fresh-eyes subagent.
- H fixed class=introduced: placeholder fragmentation after shrink
  permanently wedged later growth (split/replace require a single
  placeholder) — the fragmentation invariant above.
- H fixed class=introduced: a failed shrink truncation (reachable:
  any peer mapping the file) left coverage torn with the error
  swallowed as non-fatal — next mapped read was a process-killing
  access violation after a commit that reported success; per-site
  restore + spec'd deferral.
- M fixed class=introduced: LazyProc panic on pre-1803 windows →
  pre-flighted error.
- M fixed class=introduced ×3 (rounds 2-3): prepareShrink failures
  bypassed restore; post-truncate restore targeted the pre-shrink
  extent (structurally impossible — per-site targets now);
  single-placeholder coalesce assumption (structural counting now).
- M fixed class=introduced (round 4): mapExtent's split-succeeded/
  map-failed exit broke the invariant — unsplit on every such exit.
- L fixed class=introduced: 386 ABI hazard (build tags), alignment
  error masking (wrapped), error-path address-space leak (single
  release via the invariant; residue documented).
- L disputed class=adjacent: two pre-existing test files fail vet on
  windows/386 (untyped-constant overflow) — out-of-spec arch, no CI
  surface, production build green (reviewer accepted; user audit).
- nit fixed class=adjacent: stale read-only-pager comment (db.go).

Verified: windows/amd64 + windows/arm64 build clean, windows/amd64
vet clean, windows/386 builds to the stub; linux build+vet and full
-race suite green. Runtime verification of the windows mapping lands
with the windows CI chunk.

+++ internal/lock/proc_windows.go
@@ -33,7 +33,7 @@ const stillActive = 259
 // Awareness) — but the symbol must exist and behave correctly for
 // the build and for any future same-host classification.
 func IsAlive(pid int) bool {
-	if pid <= 0 || pid > math.MaxUint32 {
+	if pid <= 0 || uint64(pid) > math.MaxUint32 {
 		// The upper guard restores unix parity for a corrupt record's
 		// impossible pid: kill() would ESRCH, but uint32 truncation
 		// could alias a live process into false-alive.
